+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 This file documents all notable changes to the CTest Lab extension.
 
-## [Unreleased]
+## [0.3.0] — 2022-10-08
 
 ### Added
 
@@ -36,7 +36,8 @@ This file documents all notable changes to the CTest Lab extension.
 
 - Basic test discovery: [#2](https://github.com/brobeson/ctest-lab/issues/2)
 
-[unreleased]: https://github.com/brobeson/ctest-lab/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/brobeson/ctest-lab/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/brobeson/ctest-lab/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/brobeson/ctest-lab/compare/v0.1.0...v0.2.0
 [0.1.0]:
   https://github.com/brobeson/ctest-lab/compare/2e0e350936d6e22192fe289864c565795f6b7924...v0.1.0